refactor: Extract MuiCalendarPicker into a separate component
